feat: split file-selector variants in separate components
