+++ src/elements/file-selector.ts
@@ -1,1 +1,3 @@
+export * from './file-selector/file-selector-common.js';
+export * from './file-selector/file-selector-dropzone.js';
 export * from './file-selector/file-selector.js';
