+++ src/elements/file-selector/file-selector-common.ts
@@ -1,0 +1,3 @@
+export * from './common/file-selector-common.js';
+
+export { default as fileSelectorCommonStyle } from './common/file-selector-common.scss?lit&inline';
