+++ src/elements/file-selector/file-selector/file-selector.ssr.spec.ts
@@ -1,7 +1,7 @@
 import { assert } from '@open-wc/testing';
 import { html } from 'lit';
 
-import { ssrHydratedFixture } from '../core/testing/private.js';
+import { ssrHydratedFixture } from '../../core/testing/private.js';
 
 import { SbbFileSelectorElement } from './file-selector.js';
